azure_function/requirements.txt
azure-functions==1.21.3
fastapi==0.111.1
uvicorn==0.30.1
python-dotenv==1.0.1
azure-storage-blob==12.23.1
azure-servicebus==7.13.0
redis==5.2.0
openai==1.54.3
pinecone-client==5.0.1
langchain==0.3.7
langchain_community==0.3.7
rasterio==1.4.2
albumentations==1.4.21
cryptography==43.0.3
pypdf==5.1.0
